phase2-product-management/requirements.txt
# Amazon FBM Automation — Python Dependencies

# Google APIs
gspread>=5.12.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-api-python-client>=2.100.0

# Data processing
pandas>=2.1.0
openpyxl>=3.1.2
pdfplumber>=0.10.3
xlrd>=2.0.1

# Amazon SP-API
python-amazon-sp-api>=1.0.0

# Utils
python-dotenv>=1.0.0
requests>=2.31.0
